refactor: 전화번호 입력 관련 함수는 해당 컴포넌트에서 정의되도록 수정
